requirements.txt
# ── Core Django ───────────────────────────────────────────────
Django==4.2.26
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
asgiref==3.8.1

# ── Database ──────────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ── Auth / Security ───────────────────────────────────────────
django-cors-headers==4.4.0
PyJWT==2.9.0

# ── Async / WebSocket ─────────────────────────────────────────
channels==4.0.0
channels-redis==4.1.0

# ── Task Queue ────────────────────────────────────────────────
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
kombu==5.5.4
billiard==4.2.4
vine==5.1.0
amqp==5.3.1

# ── Static Files ──────────────────────────────────────────────
whitenoise==6.7.0

# ── Media / Images ────────────────────────────────────────────
Pillow==10.1.0
django-imagekit==6.1.0
pilkit==3.0

# ── Config ────────────────────────────────────────────────────
python-decouple==3.8

# ── Tagging ───────────────────────────────────────────────────
django-taggit==6.1.0

# ── AI / ML ───────────────────────────────────────────────────
openai==1.6.1

# ── Payments ──────────────────────────────────────────────────
razorpay==1.4.1

# ── Document Parsing ──────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.0
lxml>=3.1.0

# ── HTTP ──────────────────────────────────────────────────────
requests==2.32.4
httpx==0.28.1

# ── Production Server ─────────────────────────────────────────
gunicorn==23.0.0
uvicorn[standard]==0.33.0

# ── Dev / Debug ───────────────────────────────────────────────
django-debug-toolbar==4.2.0

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.5
pytest-django==4.11.1
pytest-cov==5.0.0
factory-boy==3.3.3
coverage==7.6.1
